Fix error when verifying for null values
Added an abstract class Operations which would help identify null
strings passed down from the javascript interface.

There was a bug introduced in the code, basically when a null value is
received from the javascript interface, it is actually a String object
with the "null" value loaded. But it is not clear if in any case it
returns an actual null String.