Fix error when checking for null values.
There was a problem when trying to find out whether a parameter passed
down from the cordova javascript interface was whether a null String
object or a String object containing the word "null".

This has been solved by adding a mechanism to chec for both cases, since
the documentation on apache cordova is not that clear about that matter.